requirements_bot.txt
python-telegram-bot==20.7
crewai==0.28.8
langchain-groq==0.0.1
groq==0.4.2
langchain==0.1.11
langchain-core==0.1.27
requests==2.31.0
